+++ base/feed/src/main/java/io/snaps/basefeed/data/model/MarkVideoShownRequestDto.kt
@@ -7,6 +7,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class MarkVideoShownRequestDto(
     @SerialName("videoId") val videoId: Uuid,
-    @SerialName("videoDuration") val videoDuration: Long,
-    @SerialName("duration") val duration: Long,
+    @SerialName("videoDuration") val videoDuration: Float,
+    @SerialName("duration") val duration: Float,
 )
